requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.3
aiohappyeyeballs=2.4.6=pyhd8ed1ab_0
aiohttp=3.11.13=py310hc74094e_0
aiosignal=1.3.2=pyhd8ed1ab_0
async-timeout=5.0.1=pyhd8ed1ab_1
attrs=25.1.0=pyh71513ae_0
brotli-python=1.1.0=py310hb4ad77e_2
bzip2=1.0.8=h99b78c6_7
c-ares=1.34.4=h5505292_0
ca-certificates=2025.1.31=hf0a4a13_0
cachetools=5.5.2=pyhd8ed1ab_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py310h497396d_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
click=8.1.8=pyh707e725_0
cryptography=44.0.2=py310h69cbf43_0
frozenlist=1.5.0=py310hc74094e_1
google-api-core=2.24.1=pyhd8ed1ab_0
google-auth=2.38.0=pyhd8ed1ab_0
google-cloud-core=2.4.2=pyhd8ed1ab_0
google-cloud-storage=3.1.0=pyhd8ed1ab_0
google-crc32c=1.1.2=py310h84dbace_6
google-resumable-media=2.7.2=pyhd8ed1ab_2
googleapis-common-protos=1.69.0=pyhd8ed1ab_0
grpcio=1.68.2=py310h92b281b_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
libabseil=20240722.0=cxx17_h07bc746_4
libcrc32c=1.1.2=hbdafb3b_0
libcxx=19.1.7=ha82da77_0
libffi=3.4.2=h3422bc3_5
libgrpc=1.68.2=h0a426d6_0
liblzma=5.6.4=h39f12f2_0
libprotobuf=5.28.3=h3bd63a1_1
libre2-11=2024.07.02=h07bc746_2
libsqlite=3.49.1=h3f77e49_1
libzlib=1.3.1=h8359307_2
multidict=6.1.0=py310hdf261b0_1
ncurses=6.5=h5e97a16_3
openssl=3.4.1=h81ee809_0
pip=25.0.1=pyh8b19718_0
propcache=0.2.1=py310hc74094e_1
proto-plus=1.26.0=pyhd8ed1ab_0
protobuf=5.28.3=py310h853098b_0
pyasn1=0.6.1=pyhd8ed1ab_2
pyasn1-modules=0.4.1=pyhd8ed1ab_1
pycparser=2.22=pyh29332c3_1
pyopenssl=25.0.0=pyhd8ed1ab_0
pysocks=1.7.1=pyha55dd90_7
python=3.10.16=h870587a_1_cpython
python_abi=3.10=5_cp310
pyu2f=0.1.5=pyhd8ed1ab_1
re2=2024.07.02=h6589ca4_2
readline=8.2=h1d1bf99_2
requests=2.32.3=pyhd8ed1ab_1
rsa=4.9=pyhd8ed1ab_1
setuptools=75.8.2=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025a=h78e105d_0
urllib3=2.3.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
yarl=1.18.3=py310hc74094e_1
zstandard=0.23.0=py310h2665a74_1
zstd=1.5.6=hb46c0d2_0
